Refactored sales statistics calculation
