+++ lib/bezirke_web/components/layouts/app.html.heex
@@ -2,8 +2,13 @@
   <div class="flex items-center justify-between border-b border-zinc-100 py-3 text-sm">
     <div class="flex items-center gap-4 font-semibold leading-6 text-zinc-900">
       <a href={~p"/"} class="hover:text-zinc-700">
-        Statistics
+        [Production Statistics]
       </a>
+      <a href={~p"/statistics/performances"} class="hover:text-zinc-700">
+        [Performance Statistics]
+      </a>
+    </div>
+    <div class="flex items-center gap-4 font-semibold leading-6 text-zinc-900">
       <a href={~p"/productions"} class="hover:text-zinc-700">
         Productions
       </a>
