+++ lib/bezirke_web/live/production_sales_statistics.ex
@@ -2,6 +2,7 @@ defmodule BezirkeWeb.ProductionSalesStatistics do
   use BezirkeWeb, :live_view
 
   alias Bezirke.Sales
+  alias Bezirke.Statistics
   alias Bezirke.Tour
 
   def mount(_params, _session, socket) do
@@ -80,13 +81,8 @@ defmodule BezirkeWeb.ProductionSalesStatistics do
 
     {labels, datasets} =
       production_statistics
-      |> build_chart()
-
-    datasets =
-      datasets
-      |> Enum.map(fn {label, tickets_count} ->
-        %{label: label, data: tickets_count}
-      end)
+      |> Enum.map(fn {production, sales_figures, _, _} -> {production, sales_figures} end)
+      |> Statistics.build_chart()
 
     {production_statistics, labels, datasets}
   end
@@ -121,95 +117,5 @@ defmodule BezirkeWeb.ProductionSalesStatistics do
       capacity,
       tickets_count,
     }
-  end
-
-  defp build_chart([]), do: {[], []}
-
-  defp build_chart(production_statistics) do
-    labels = get_labels(production_statistics)
-
-    datasets = build_datasets(labels, production_statistics)
-
-    {labels, datasets}
-  end
-
-  defp get_labels(production_statistics) do
-    production_statistics
-    |> Enum.flat_map(fn {_, sales_figures, _, _} ->
-      Enum.map(sales_figures, &(DateTime.to_date(&1.record_date)))
-    end)
-    |> Enum.sort_by(&(&1), Date)
-    |> get_labels_from_sales_figures()
-  end
-
-  defp get_labels_from_sales_figures([]), do: []
-
-  defp get_labels_from_sales_figures(sales_figures) do
-    start_date =
-      sales_figures
-      |> List.first()
-      |> Date.add(-1)
-
-    end_date =
-      sales_figures
-      |> List.last()
-      |> Date.add(1)
-
-    build_labels(start_date, end_date, [])
-    |> Enum.reverse()
-  end
-
-  defp build_labels(current_date, end_date, labels) do
-    case Date.compare(current_date, end_date) do
-      :eq -> [current_date | labels]
-      _ -> build_labels(Date.add(current_date, 1), end_date, [current_date | labels])
-    end
-  end
-
-  defp build_datasets(labels, production_statistics) do
-    production_statistics
-    |> Enum.map(fn {production, sales_figures, _, _} ->
-      dataset =
-        sales_figures
-        |> Enum.sort_by(&(&1.record_date), DateTime)
-        |> build_dataset(labels, [])
-        |> Enum.reverse()
-
-      {production, dataset}
-    end)
-  end
-
-  defp build_dataset(_, [], tickets_count), do: tickets_count
-
-  defp build_dataset([], [_ | next_dates], [latest_tickets_count | _] = tickets_count) do
-    build_dataset([], next_dates, [latest_tickets_count | tickets_count])
-  end
-
-  defp build_dataset(sales_figures, [current_date | next_dates], []) do
-    {total_tickets_count, sales_figures} = sum_tickets_count(0, current_date, sales_figures)
-
-    build_dataset(sales_figures, next_dates, [total_tickets_count])
-  end
-
-  defp build_dataset(sales_figures, [current_date | next_dates], [latest_tickets_count | _] = tickets_count) do
-    {total_tickets_count, sales_figures} = sum_tickets_count(latest_tickets_count, current_date, sales_figures)
-
-    build_dataset(sales_figures, next_dates, [total_tickets_count | tickets_count])
-  end
-
-  defp sum_tickets_count(total_tickets_count, _, []), do: {total_tickets_count, []}
-
-  defp sum_tickets_count(
-    total_tickets_count,
-    current_date,
-    [
-      %Bezirke.Sales.SalesFigures{record_date: record_date, tickets_count: tickets_count}
-      | next_sales_figures
-    ] = sales_figures
-  ) do
-    case Date.compare(current_date, DateTime.to_date(record_date)) do
-      :eq -> sum_tickets_count(tickets_count + total_tickets_count, current_date, next_sales_figures)
-      _ -> {total_tickets_count, sales_figures}
-    end
   end
 end
